backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

langgraph==0.2.34
langchain==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.2

pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.9.2